chat view updated (devInfo not done yet
