--- app/src/main/java/com/anpha/android_huflit/chatMenu.java
@@ -1,44 +1,0 @@
-package com.anpha.android_huflit;
-
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.AppCompatActivity;
-
-import androidx.annotation.NonNull;
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.appcompat.widget.Toolbar;
-
-import android.content.Intent;
-import android.graphics.drawable.Drawable;
-import android.os.Bundle;
-import android.view.Menu;
-import android.view.MenuInflater;
-import android.view.MenuItem;
-import android.widget.TextView;
-
-
-public class chatMenu extends AppCompatActivity {
-
-    TextView txtUser;
-    Drawable boy;
-
-    Toolbar toolbarChatMenu;
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_chat_menu);
-
-        toolbarChatMenu = findViewById(R.id.toolbarChatMenu);
-        toolbarChatMenu.setOnMenuItemClickListener(new Toolbar.OnMenuItemClickListener() {
-            @Override
-            public boolean onMenuItemClick(MenuItem item) {
-                if(item.getItemId() == R.id.iconMenu){
-                    finish();
-                }
-                return false;
-            }
-        });
-
-    }
-
-
-}
