+++ app/src/main/java/com/anpha/android_huflit/textChat.java
@@ -1,16 +1,25 @@
 package com.anpha.android_huflit;
-
+import androidx.appcompat.app.ActionBarDrawerToggle;
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
 import androidx.appcompat.widget.Toolbar;
-
+import androidx.drawerlayout.widget.DrawerLayout;
+import android.widget.LinearLayout;
 import android.content.Intent;
 import android.os.Bundle;
 import android.view.Menu;
 import android.view.MenuInflater;
 import android.view.MenuItem;
+import androidx.core.view.GravityCompat;
+import android.widget.PopupWindow;
+import android.view.View;
+import android.view.LayoutInflater;
+import android.view.Gravity;
+import android.widget.ScrollView;
 
 public class textChat extends AppCompatActivity {
+
+    PopupWindow popupWindow;
 
     Toolbar toolbarChat;
     @Override
@@ -19,16 +28,25 @@ public class textChat extends AppCompatActivity {
         setContentView(R.layout.activity_text_chat);
 
         toolbarChat = findViewById(R.id.toolbarChat);
+
+        View popupView = LayoutInflater.from(this).inflate(R.layout.popup_chat_menu, null);
+        popupWindow = new PopupWindow(
+                popupView,
+                LinearLayout.LayoutParams.MATCH_PARENT,
+                1700,
+                true
+        );
+
         toolbarChat.setOnMenuItemClickListener(new Toolbar.OnMenuItemClickListener() {
             @Override
             public boolean onMenuItemClick(MenuItem item) {
-                if(item.getItemId() == R.id.iconMenu){
-                    Intent i = new Intent(textChat.this,chatMenu.class);
-                    startActivity(i);
+                if(item.getItemId() == R.id.iconMenu)
+                {
+                    popupWindow.showAsDropDown(toolbarChat.findViewById(R.id.iconMenu));
+                    return true;
                 }
                 return false;
             }
         });
-
     }
 }
