+++ app/src/main/res/values/styles.xml
@@ -3,5 +3,6 @@
 <style name="AppTheme" parent="Theme.AppCompat.Light.NoActionBar">
     <item name="android:layoutDirection">ltr</item>
     <item name="android:layout_gravity">start</item>
+    <item name="android:windowFullscreen">true</item>
 </style>
 </resources>
